+++ src/components/RelativeRangeDatePicker/components/Control/i18n/en.json
@@ -1,0 +1,3 @@
+{
+  "Range date picker": "Range date picker"
+}
