+++ src/components/RelativeRangeDatePicker/components/Control/i18n/ru.json
@@ -1,0 +1,3 @@
+{
+  "Range date picker": "Выбор диапазона дат"
+}
