+++ src/components/RelativeRangeDatePicker/i18n/en.json
@@ -4,6 +4,5 @@
   "\"To\"": "\"To\"",
   "\"To\" is required.": "\"To\" is required.",
   "\"From\" can't be after \"To\".": "\"From\" can't be after \"To\".",
-  "to": "to",
-  "Range date picker": "Range date picker"
+  "to": "to"
 }
