+++ src/components/RelativeRangeDatePicker/i18n/ru.json
@@ -4,6 +4,5 @@
   "\"To\"": "«До»",
   "\"To\" is required.": "«До» обязательно.",
   "\"From\" can't be after \"To\".": "«От» не может быть позже чем «До».",
-  "to": "до",
-  "Range date picker": "Выбор диапазона дат"
+  "to": "до"
 }
